Add files via upload
**German Verbs with Prepositions v1.0**
This is the first official release of the German Verbs with Prepositions deck for Anki. It includes over 292 verbs each with an example sentence and their translations in English and Persian. This release aims to assist learners in understanding and using German prepositions more effectively.

**Features:**
- Comprehensive data on German verbs and prepositions
- Example sentences in German with English and Persian translations
- Easy to import CSV and Excel files for Anki users

**Installation Instructions:**
Download the `.apkg` file for Anki or the `.csv`/`.xlsx` files to import the deck into your Anki application manually.